api/requirements-worker.txt
boto3==1.34.145
pillow==10.4.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
torch==2.3.1
torchvision==0.18.1
open-clip-torch==2.26.1
